# cmd/api/docs/swagger.yaml
definitions:
...
  requests.Coupon:
    properties:
      code:
        type: string
      discount_percent:
        type: number
      expiry_date:
        type: string
      maximum_discount_price:
        type: number
      minimum_purchase_price:
        type: number
      usage_limits:
        type: integer
    required:
    - code
    - discount_percent
    - expiry_date
    - maximum_discount_price
    - minimum_purchase_price
    - usage_limits
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/coupon/AddCoupons:
    post:
      description: Admin can add coupon
      operationId: AddCoupon
      parameters:
      - description: Input true info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.Coupon'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: api for add Coupons for ecommerce
      tags:
      - Coupon
  /admin/coupon/Delete/{CouponID}:
    delete:
      consumes:
      - application/json
      description: Admin can delete a coupon
      operationId: delete-coupon
      parameters:
      - description: CouponID
        in: path
        name: CouponID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can delete a coupon
      tags:
      - Coupon
  /admin/coupon/Update/{CouponID}:
    patch:
      consumes:
      - application/json
      description: Admin can update existing coupon
      operationId: update-coupon
      parameters:
      - description: CouponID
        in: path
        name: CouponID
        required: true
        type: integer
      - description: details of coupon to be updated
        in: body
        name: coupon_details
        required: true
        schema:
          $ref: '#/definitions/requests.Coupon'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin can update existing coupon
      tags:
      - Coupon
  /admin/coupon/Viewcoupon/{id}:
    get:
      consumes:
      - application/json
      description: Admins can see Coupons with coupon_id
      operationId: find-Coupon-by-id
      parameters:
      - description: CouponID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admins can see Coupons with coupon_id
      tags:
      - Coupon
  /admin/coupon/couponlist:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all coupons
      operationId: List-all-coupons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all coupons
      tags:
      - Coupon
...
  /coupon/apply:
    patch:
      consumes:
      - application/json
      description: User can apply coupon to the cart
      operationId: apply-coupon-to-cart
      parameters:
      - description: Coupon code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User can apply a coupon to the cart
      tags:
      - Cart
  /coupon/coupons:
    get:
      consumes:
      - application/json
      description: Endpoint for getting all coupons in user side
      operationId: List-all-coupons-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all coupons for users
      tags:
      - Coupon
...
swagger: "2.0"